requirements.txt
streamlit
mysql.connector
uuid
streamlit_calendar
smtplib
email.message
ssl
json